auth/requirements.txt
flask==3.1.0
pyjwt[crypto]==2.10.1
requests==2.32.3
flask-limiter==3.8.0
bcrypt==4.2.0
gunicorn==24.1.1
redis==5.0.8
flask-cors==5.0.0
